pyproject.toml
[project]
name = "drf-action-serializers"
version = "0.1.0"
description = "Easily specify a serializer for each action and method combination"
authors = [
    {name = "Kevin Renskers", email = "kevin@loopwerk.io"},
]
license = "MIT"
license-files = [ "LICENSE" ]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "djangorestframework>=3.14",
]

[dependency-groups]
dev = [
    "pytest>=8.3.5",
    "pytest-django>=4.10.0",
    "ruff>=0.11.2",
    "drf-spectacular>=0.26",
]

[project.optional-dependencies]
spectacular = ["drf-spectacular>=0.26"]

[project.urls]
Homepage = "https://github.com/loopwerk/drf-action-serializers"
Repository = "https://github.com/loopwerk/drf-action-serializers.git"
Issues = "https://github.com/loopwerk/drf-action-serializers/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
package = true

[tool.ruff]
line-length = 120
lint.extend-select = ["I", "N"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "tests.settings"
python_files = "test*.py"
filterwarnings = ["ignore::DeprecationWarning"]